// src/commons/types/generated/types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: any;
  Upload: any;
};

export type IArtist = {
  __typename?: 'Artist';
  active_name: Scalars['String'];
  artistImageURL: Scalars['String'];
  category: ICategory;
  description: Scalars['String'];
  id: Scalars['String'];
  pick_user: Array<ILikeArtist>;
  promotion_url?: Maybe<Scalars['String']>;
};

export type IBoardAddress = {
  __typename?: 'BoardAddress';
  address: Scalars['String'];
  address_city: Scalars['String'];
  address_district: Scalars['String'];
  id: Scalars['String'];
  lat: Scalars['Float'];
  lng: Scalars['Float'];
};

export type IBoardAddressInput = {
  address: Scalars['String'];
  lat: Scalars['Float'];
  lng: Scalars['Float'];
};

export type IBoardImages = {
  __typename?: 'BoardImages';
  boards: IBoards;
  id: Scalars['String'];
  url: Scalars['String'];
};

export type IBoards = {
  __typename?: 'Boards';
  artist: IArtist;
  boardAddress: IBoardAddress;
  boardImageURL: Array<IBoardImages>;
  category?: Maybe<ICategory>;
  comments: Array<IComments>;
  contents: Scalars['String'];
  createAt: Scalars['DateTime'];
  end_time?: Maybe<Scalars['DateTime']>;
  id: Scalars['String'];
  isShowTime?: Maybe<Scalars['Boolean']>;
  start_time?: Maybe<Scalars['DateTime']>;
  title: Scalars['String'];
  updatedAt: Scalars['DateTime'];
};

export type ICategory = {
  __typename?: 'Category';
  id: Scalars['String'];
  name: Scalars['String'];
};

export type ICity = {
  __typename?: 'City';
  district: Array<IDistrict>;
  id: Scalars['String'];
  name: Scalars['String'];
};

export type IComments = {
  __typename?: 'Comments';
  board: IBoards;
  content: Scalars['String'];
  createAt: Scalars['DateTime'];
  id: Scalars['String'];
  user: IUser;
};
// ...
export type IDistrict = {
  __typename?: 'District';
  city: ICity;
  district: Scalars['String'];
  id: Scalars['String'];
};

export type IFetchDistricts = {
  __typename?: 'FetchDistricts';
  district: Array<Scalars['String']>;
  name: Scalars['String'];
};

export type ILikeArtist = {
  __typename?: 'LikeArtist';
  artist: IArtist;
  id: Scalars['String'];
  user: IUser;
};

export type IMember = {
  __typename?: 'Member';
  artist?: Maybe<IArtist>;
  id: Scalars['String'];
  memberImageURL: Scalars['String'];
  name: Scalars['String'];
  role: Scalars['String'];
};
// ...
export type IQuery = {
  __typename?: 'Query';
  districtList: Array<IDistrict>;
  fetchArtist: IArtist;
  fetchArtistCount: Scalars['Int'];
  fetchArtistWithoutAuth: IArtist;
  fetchBoard: IBoards;
  fetchBoardsBySearch: Array<IBoards>;
  fetchByComment: IComments;
  fetchCategories: Array<ICategory>;
  fetchCity: IFetchDistricts;
  fetchCitys: Array<ICity>;
  fetchComments: Array<IComments>;
  fetchMapBoards: Array<IBoards>;
  fetchMembers: Array<IMember>;
  fetchRecentBoards: Array<IBoards>;
  fetchUser: IUser;
};


export type IQueryFetchArtistCountArgs = {
  artistId: Scalars['String'];
};


export type IQueryFetchArtistWithoutAuthArgs = {
  artistId: Scalars['String'];
};


export type IQueryFetchBoardArgs = {
  boardId: Scalars['String'];
};


export type IQueryFetchBoardsBySearchArgs = {
  categoryId?: InputMaybe<Array<Scalars['String']>>;
  districtId?: InputMaybe<Scalars['String']>;
  page?: InputMaybe<Scalars['Int']>;
};
// ...
export type IUser = {
  __typename?: 'User';
  authorities: IUserAuthority;
  email: Scalars['String'];
  id: Scalars['String'];
  liked_artist: Array<ILikeArtist>;
  nickname: Scalars['String'];
  userImageURL: Scalars['String'];
  wrong_pass: Scalars['Int'];
};

export type IUserAuthority = {
  __typename?: 'UserAuthority';
  artist: IArtist;
  artistId: Scalars['String'];
  authority: Scalars['String'];
  id: Scalars['String'];
  userId: Scalars['String'];
};